requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0

# Data sources
yfinance>=0.2.0
tavily-python>=0.5.0
langchain-tavily>=0.2.0

# Google Sheets
gspread>=6.0.0
oauth2client>=4.1.3
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0

# Environment and utilities
python-dotenv>=1.0.0
tenacity>=8.0.0
